♻️  calls screen component
