+++ package.json
@@ -21,6 +21,7 @@
     "eslint": "^5.5.0",
     "eslint-plugin-jest": "^21.22.0",
     "jest": "^23.5.0",
+    "react-dom": "^16.5.0",
     "react-test-renderer": "16.5.0",
     "regenerator-runtime": "^0.12.1"
   }
